+++ src/modules/scheduler/router.py
@@ -1,377 +1,13 @@
-import json, os
-import platform
-from typing import List
-import datetime
-from fastapi import APIRouter, Depends, HTTPException, status, Path, Query
-from sqlalchemy.orm import Session
-from apscheduler.jobstores.base import JobLookupError
+from fastapi import APIRouter
 
-from core.database import get_db
-from modules.scheduler import models, schemas, loader
-from modules.scheduler.service import job_definition_service
-from modules.scheduler import scheduler_instance, service
-from util import logger_util, config_util, time_util
-
-logger = logger_util.get_logger(__name__)
+from modules.scheduler.routers import system, dashboard, filesystem, jobs, workflows, scheduler, config
 
 router = APIRouter(prefix="/api")
 
-#
-# --- System Endpoints ---
-#
-@router.get("/system/os", tags=["System"], summary="Get OS Information")
-def get_os_info():
-    """Returns the operating system type (e.g., 'Windows', 'Linux')."""
-    return {"os_type": platform.system()}
-
-#
-# --- Dashboard Endpoints ---
-#
-@router.get("/dashboard/summary", response_model=schemas.DashboardSummary, tags=["Dashboard"], summary="Get Dashboard Summary", description="Provides a high-level summary of job statuses.")
-def get_dashboard_summary(db: Session = Depends(get_db)):
-    try:
-        return service.get_dashboard_summary(db)
-    except Exception as e:
-        logger.error(f"Error fetching dashboard summary: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to fetch dashboard summary")
-
-@router.get("/logs", response_model=List[schemas.ProcessExecutionLogInfo], tags=["Dashboard"], summary="Get Execution Logs", description="Retrieves a paginated list of job execution logs.")
-def get_execution_logs(skip: int = Query(0, ge=0), limit: int = Query(100, ge=1, le=200), db: Session = Depends(get_db)):
-    try:
-        return service.get_execution_logs(db, skip=skip, limit=limit)
-    except Exception as e:
-        logger.error(f"Error fetching execution logs: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to fetch execution logs")
-
-@router.get("/timeline-items", response_model=List[schemas.TimelineItem], tags=["Dashboard"], summary="Get Timeline Data", description="Provides data for the job execution timeline, including scheduled and historical runs.")
-def get_timeline_data(db: Session = Depends(get_db)):
-    try:
-        return service.get_timeline_data(db)
-    except Exception as e:
-        logger.error(f"Error fetching timeline data: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.get("/unified-jobs", response_model=List[schemas.UnifiedJobItem], tags=["Dashboard"])
-def get_unified_jobs(db: Session = Depends(get_db)):
-    """
-    Retrieves a unified list of all jobs and workflows.
-    """
-    try:
-        return service.get_unified_jobs_list(db)
-    except Exception as e:
-        logger.error(f"Error fetching unified jobs list: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to fetch unified jobs list")
-
-#
-# --- Filesystem Endpoints ---
-#
-@router.get("/filesystem/list-dirs", response_model=List[str], tags=["Filesystem"], summary="List Subdirectories in Work Directory")
-def list_work_dir_subdirectories(path: str = Query("", description="The relative path within the work directory to scan.")):
-    """
-    Lists subdirectories within the configured scheduler work_dir.
-    This is useful for providing autocompletion for the 'cwd' field in a UI.
-    """
-    try:
-        return service.list_subdirectories(relative_path=path)
-    except Exception as e:
-        logger.error(f"Error listing subdirectories for path '{path}': {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to list directories")
-
-# --- Job Definition Endpoints ---
-#
-@router.get("/available-tasks", response_model=List[schemas.AvailableTask], tags=["Job Definitions"], summary="List Available Tasks")
-def get_available_tasks():
-    """
-    Returns a list of all available task types (Python, Shell, etc.)
-    that can be used to create jobs, including their parameters.
-    """
-    try:
-        return service.get_available_tasks()
-    except Exception as e:
-        logger.error(f"Error fetching available tasks: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to fetch available tasks")
-
-
-@router.get("/jobs", response_model=List[schemas.Job], tags=["Job Definitions"], summary="List All Job Definitions")
-def read_jobs(db: Session = Depends(get_db), skip: int = Query(0, ge=0), limit: int = Query(100, ge=1, le=500)):
-    jobs = job_definition_service.get_multi(db, skip=skip, limit=limit)
-    return [schemas.Job.model_validate(job) for job in jobs]
-
-@router.post("/jobs", response_model=schemas.Job, status_code=status.HTTP_201_CREATED, tags=["Job Definitions"], summary="Create a New Job Definition")
-def create_job(job_in: schemas.JobCreate, db: Session = Depends(get_db)):
-    """
-    Creates a new job definition in the database and adds it to the scheduler.
-    The input is validated against the `JobCreate` schema, which uses a discriminated
-    union to validate `task_parameters` based on `task_type`.
-    """
-    # The service layer will handle the creation from the new schema.
-    db_job = service.create_job_from_schema(db, job_in=job_in)
-    if not db_job:
-        # The service should handle ID conflicts and raise an exception or return None.
-        raise HTTPException(status_code=409, detail="Job with this ID might already exist or creation failed.")
-    loader.sync_jobs_from_db()
-    return schemas.Job.model_validate(db_job)
-
-@router.get("/jobs/{job_id}", response_model=schemas.Job, tags=["Job Definitions"])
-def read_job(job_id: str, db: Session = Depends(get_db)):
-    db_job = job_definition_service.get(db, id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    return schemas.Job.model_validate(db_job)
-
-@router.put("/jobs/{job_id}", response_model=schemas.Job, tags=["Job Definitions"])
-def update_job(job_id: str, job_in: schemas.JobUpdate, db: Session = Depends(get_db)):
-    db_job = job_definition_service.get(db, id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    # The service layer will need to handle the update from the new schema.
-    db_job = service.update_job_from_schema(db, db_obj=db_job, job_in=job_in)
-    loader.sync_jobs_from_db()
-    return schemas.Job.model_validate(db_job)
-
-@router.delete("/jobs/{job_id}", status_code=status.HTTP_204_NO_CONTENT, tags=["Job Definitions"])
-def delete_job(job_id: str, db: Session = Depends(get_db)):
-    db_job = job_definition_service.remove(db, id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    
-    try:
-        scheduler_instance.scheduler.remove_job(job_id)
-        logger.info(f"Removed job '{job_id}' from scheduler.")
-    except JobLookupError:
-        logger.warning(f"Job '{job_id}' was not found in the scheduler for removal, but was deleted from the database.")
-
-    return Response(status_code=status.HTTP_204_NO_CONTENT)
-
-# --- Workflow Definition Endpoints ---
-
-@router.post("/workflows", response_model=schemas.Workflow, status_code=status.HTTP_201_CREATED, tags=["Workflow Definitions"])
-def create_workflow(workflow_in: schemas.WorkflowCreate, db: Session = Depends(get_db)):
-    db_workflow = service.workflow_service.create_with_steps(db, obj_in=workflow_in)
-    loader.schedule_workflow(db_workflow)
-    return db_workflow
-
-@router.get("/workflows", response_model=List[schemas.Workflow], tags=["Workflow Definitions"])
-def read_workflows(db: Session = Depends(get_db), skip: int = 0, limit: int = 100):
-    workflows = service.workflow_service.get_multi(db, skip=skip, limit=limit)
-    return workflows
-
-@router.get("/workflows/{workflow_id}", response_model=schemas.Workflow, tags=["Workflow Definitions"])
-def read_workflow(workflow_id: int, db: Session = Depends(get_db)):
-    db_workflow = service.workflow_service.get(db, id=workflow_id)
-    if db_workflow is None:
-        raise HTTPException(status_code=404, detail="Workflow not found")
-    return db_workflow
-
-@router.put("/workflows/{workflow_id}", response_model=schemas.Workflow, tags=["Workflow Definitions"])
-def update_workflow(workflow_id: int, workflow_in: schemas.WorkflowCreate, db: Session = Depends(get_db)):
-    db_workflow = service.workflow_service.get(db, id=workflow_id)
-    if db_workflow is None:
-        raise HTTPException(status_code=404, detail="Workflow not found")
-    db_workflow = service.workflow_service.update_with_steps(db, db_obj=db_workflow, obj_in=workflow_in)
-    loader.schedule_workflow(db_workflow)
-    return db_workflow
-
-@router.delete("/workflows/{workflow_id}", status_code=status.HTTP_204_NO_CONTENT, tags=["Workflow Definitions"])
-def delete_workflow(workflow_id: int, db: Session = Depends(get_db)):
-    db_workflow = service.workflow_service.remove(db, id=workflow_id)
-    if db_workflow is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    loader.remove_workflow_job(workflow_id)
-    return
-
-# --- Workflow Control Endpoints ---
-
-@router.post("/workflows/{workflow_id}/pause", tags=["Workflow Control"], summary="Pause a Workflow")
-def pause_workflow(workflow_id: int, db: Session = Depends(get_db)):
-    """
-    Pauses a workflow by removing its job from the scheduler and marking it as disabled.
-    """
-    service.update_workflow_enabled_status(db, workflow_id=workflow_id, is_enabled=False)
-    try:
-        loader.remove_workflow_job(workflow_id)
-        return {"message": f"Workflow '{workflow_id}' paused successfully."}
-    except Exception as e:
-        logger.error(f"Error removing workflow job {workflow_id} from scheduler: {e}", exc_info=True)
-        # Still return success as the DB state is updated
-        return {"message": f"Workflow '{workflow_id}' marked as disabled, but could not be removed from scheduler."}
-
-
-@router.post("/workflows/{workflow_id}/resume", tags=["Workflow Control"], summary="Resume a Workflow")
-def resume_workflow(workflow_id: int, db: Session = Depends(get_db)):
-    """
-    Resumes a workflow by scheduling it and marking it as enabled.
-    """
-    db_workflow = service.update_workflow_enabled_status(db, workflow_id=workflow_id, is_enabled=True)
-    if not db_workflow:
-        raise HTTPException(status_code=404, detail="Workflow not found")
-
-    try:
-        loader.schedule_workflow(db_workflow)
-        return {"message": f"Workflow '{db_workflow.name}' resumed successfully."}
-    except Exception as e:
-        logger.error(f"Error scheduling workflow {workflow_id}: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail=f"Failed to schedule workflow {workflow_id}.")
-
-
-# --- Scheduler Control Endpoints ---
-@router.get("/scheduler/jobs", response_model=List[schemas.Job], tags=["Scheduler Control"])
-def get_scheduled_jobs():
-    try:
-        # This service function will need to be updated to return data
-        # that can be validated by the new schemas.Job model.
-        return service.get_scheduled_jobs_info()
-    except Exception as e:
-        logger.error(f"Error fetching scheduled jobs: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to fetch scheduled jobs")
-
-# --- job edit Endpoints ---
-
-@router.post("/scheduler/jobs/{job_id}/pause", tags=["Scheduler Control"])
-def pause_scheduled_job(job_id: str):
-    try:
-        scheduler_instance.scheduler.pause_job(job_id)
-        return {"message": f"Job '{job_id}' paused successfully."}
-    except JobLookupError:
-        raise HTTPException(status_code=404, detail=f"Job '{job_id}' not found.")
-
-@router.post("/scheduler/jobs/{job_id}/resume", tags=["Scheduler Control"])
-def resume_scheduled_job(job_id: str):
-    try:
-        scheduler_instance.scheduler.resume_job(job_id)
-        return {"message": f"Job '{job_id}' resumed successfully."}
-    except JobLookupError:
-        raise HTTPException(status_code=404, detail=f"Job '{job_id}' not found.")
-
-@router.post("/scheduler/jobs/{job_id}/run", tags=["Scheduler Control"])
-def run_scheduled_job_immediately(job_id: str):
-    try:
-        scheduler_instance.scheduler.modify_job(job_id, next_run_time=time_util.get_current_utc_time())
-        return {"message": f"Job '{job_id}' scheduled for immediate execution."}
-    except JobLookupError:
-        raise HTTPException(status_code=404, detail=f"Job '{job_id}' not found.")
-
-
-
-
-@router.post("/jobs/bulk/delete", status_code=status.HTTP_200_OK, tags=["Job Definitions"])
-def delete_bulk_jobs(payload: schemas.BulkJobUpdate, db: Session = Depends(get_db)):
-    job_ids = payload.job_ids
-    if not job_ids:
-        raise HTTPException(status_code=400, detail="No job IDs provided")
-    
-    try:
-        deleted_count = service.delete_bulk_jobs(db, job_ids=job_ids)
-        if deleted_count > 0:
-            loader.sync_jobs_from_db()
-        return {"message": f"Successfully deleted {deleted_count} jobs."}
-    except Exception as e:
-        logger.error(f"Error during bulk deletion of jobs: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to delete jobs")
-
-@router.post("/scheduler/jobs/bulk/pause", tags=["Scheduler Control"])
-def pause_bulk_scheduled_jobs(payload: schemas.BulkJobUpdate):
-    if not payload.job_ids:
-        raise HTTPException(status_code=400, detail="No job IDs provided")
-    try:
-        result = service.pause_bulk_scheduled_jobs(payload.job_ids)
-        if result["failed"]:
-            return {"message": "Partial success", "paused": result["paused"], "failed": result["failed"]}
-        return {"message": "All selected jobs paused successfully."}
-    except Exception as e:
-        logger.error(f"Error during bulk pause of jobs: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to pause jobs")
-
-@router.post("/scheduler/jobs/bulk/resume", tags=["Scheduler Control"])
-def resume_bulk_scheduled_jobs(payload: schemas.BulkJobUpdate):
-    if not payload.job_ids:
-        raise HTTPException(status_code=400, detail="No job IDs provided")
-    try:
-        result = service.resume_bulk_scheduled_jobs(payload.job_ids)
-        if result["failed"]:
-            return {"message": "Partial success", "resumed": result["resumed"], "failed": result["failed"]}
-        return {"message": "All selected jobs resumed successfully."}
-    except Exception as e:
-        logger.error(f"Error during bulk resume of jobs: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to resume jobs")
-
-@router.get("/jobs/{job_id}/history", response_model=List[schemas.ProcessExecutionLogInfo], tags=["Job Details"])
-def get_job_execution_history(job_id: str, db: Session = Depends(get_db)):
-    try:
-        return service.get_job_execution_history(db, job_id=job_id)
-    except Exception as e:
-        logger.error(f"Error fetching job execution history for job {job_id}: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to fetch job execution history")
-
-
-
-
-@router.get("/jobs_yaml", tags=["Configuration"])
-def get_jobs_yaml_content():
-    try:
-        content = config_util.read_jobs_yaml_content()
-        return {"content": content}
-    except FileNotFoundError as e:
-        logger.error(f"Error reading jobs.yaml: {e}", exc_info=True)
-        raise HTTPException(status_code=404, detail=str(e))
-    except Exception as e:
-        logger.error(f"Error reading jobs.yaml: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="Failed to read jobs.yaml")
-
-@router.get("/settings/notifications", tags=["Settings"])
-def get_notification_settings():
-    try:
-        settings = config_util.get_notification_settings()
-        return settings
-    except IOError as e:
-        logger.error(f"Error reading notification settings: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail=f"Failed to read notification settings: {e}")
-    except Exception as e:
-        logger.error(f"An unexpected error occurred while reading notification settings: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="An unexpected error occurred while reading notification settings.")
-
-@router.post("/settings/notifications", tags=["Settings"])
-def update_notification_settings(email_recipients: str = "", webhook_url: str = ""):
-    settings = {
-        "email_recipients": email_recipients,
-        "webhook_url": webhook_url
-    }
-    try:
-        config_util.update_notification_settings(settings)
-        return {"message": "Notification settings updated successfully."}
-    except IOError as e:
-        logger.error(f"Error writing notification settings: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail=f"Failed to update notification settings: {e}")
-    except Exception as e:
-        logger.error(f"An unexpected error occurred while writing notification settings: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail="An unexpected error occurred while updating notification settings.")
-
-@router.post("/workflows/{workflow_id}/run", tags=["Workflow Control"], summary="Run a Workflow Immediately")
-def run_workflow_immediately(
-    workflow_id: int,
-    run_config: schemas.WorkflowRunCreate,
-    db: Session = Depends(get_db)
-):
-    """
-    Triggers an immediate, one-off execution of a workflow with specified runtime parameters.
-    """
-    db_workflow = service.workflow_service.get(db, id=workflow_id)
-    if not db_workflow:
-        raise HTTPException(status_code=404, detail="Workflow not found")
-
-    try:
-        result = service.run_workflow_immediately(db, workflow_id=workflow_id, params=run_config.params_val)
-        return result
-    except Exception as e:
-        logger.error(f"Error triggering immediate run for workflow {workflow_id}: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail=f"Failed to trigger workflow {workflow_id}.")
-
-@router.get("/workflow-runs/{run_id}/logs", response_model=List[schemas.ProcessExecutionLogInfo], tags=["Workflow Runs"])
-def get_workflow_run_logs(run_id: int, db: Session = Depends(get_db)):
-    """
-    Retrieves all execution logs for a specific workflow run.
-    """
-    logs = db.query(models.ProcessExecutionLog).filter(models.ProcessExecutionLog.workflow_run_id == run_id).order_by(models.ProcessExecutionLog.start_time).all()
-    if not logs:
-        raise HTTPException(status_code=404, detail="Logs for this workflow run not found")
-    return logs
+router.include_router(system.router)
+router.include_router(dashboard.router)
+router.include_router(filesystem.router)
+router.include_router(jobs.router)
+router.include_router(workflows.router)
+router.include_router(scheduler.router)
+router.include_router(config.router)
